core: split up mavlink command sender and receiver
This was straight forward and long overdue.
