+++ src/mavsdk/core/mavlink_command_sender.cpp
@@ -1,4 +1,4 @@
-#include "mavlink_commands.h"
+#include "mavlink_command_sender.h"
 #include "system_impl.h"
 #include <cmath>
 #include <future>
@@ -435,132 +435,6 @@ float MavlinkCommandSender::maybe_reserved(const std::optional<float>& maybe_par
             return 0.0f;
         } else {
             return NAN;
-        }
-    }
-}
-
-MavlinkCommandReceiver::MavlinkCommandReceiver(SystemImpl& system_impl) : _parent(system_impl)
-{
-    _parent.register_mavlink_message_handler(
-        MAVLINK_MSG_ID_COMMAND_LONG,
-        [this](const mavlink_message_t& message) { receive_command_long(message); },
-        this);
-
-    _parent.register_mavlink_message_handler(
-        MAVLINK_MSG_ID_COMMAND_INT,
-        [this](const mavlink_message_t& message) { receive_command_int(message); },
-        this);
-}
-
-MavlinkCommandReceiver::~MavlinkCommandReceiver()
-{
-    unregister_all_mavlink_command_handlers(this);
-
-    _parent.unregister_all_mavlink_message_handlers(this);
-}
-
-void MavlinkCommandReceiver::receive_command_int(const mavlink_message_t& message)
-{
-    MavlinkCommandReceiver::CommandInt cmd(message);
-
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    for (auto& handler : _mavlink_command_int_handler_table) {
-        if (handler.cmd_id == cmd.command) {
-            // The client side can pack a COMMAND_ACK as a response to receiving the command.
-            auto maybe_message = handler.callback(cmd);
-            if (maybe_message) {
-                _parent.send_message(maybe_message.value());
-            }
-        }
-    }
-}
-
-void MavlinkCommandReceiver::receive_command_long(const mavlink_message_t& message)
-{
-    MavlinkCommandReceiver::CommandLong cmd(message);
-
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    for (auto& handler : _mavlink_command_long_handler_table) {
-        if (handler.cmd_id == cmd.command) {
-            // The client side can pack a COMMAND_ACK as a response to receiving the command.
-            auto maybe_message = handler.callback(cmd);
-            if (maybe_message) {
-                _parent.send_message(maybe_message.value());
-            }
-        }
-    }
-}
-
-void MavlinkCommandReceiver::register_mavlink_command_handler(
-    uint16_t cmd_id, const MavlinkCommandIntHandler& callback, const void* cookie)
-{
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    MAVLinkCommandIntHandlerTableEntry entry = {cmd_id, callback, cookie};
-    _mavlink_command_int_handler_table.push_back(entry);
-}
-
-void MavlinkCommandReceiver::register_mavlink_command_handler(
-    uint16_t cmd_id, const MavlinkCommandLongHandler& callback, const void* cookie)
-{
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    MAVLinkCommandLongHandlerTableEntry entry = {cmd_id, callback, cookie};
-    _mavlink_command_long_handler_table.push_back(entry);
-}
-
-void MavlinkCommandReceiver::unregister_mavlink_command_handler(uint16_t cmd_id, const void* cookie)
-{
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    // COMMAND_INT
-    for (auto it = _mavlink_command_int_handler_table.begin();
-         it != _mavlink_command_int_handler_table.end();
-         /* no ++it */) {
-        if (it->cmd_id == cmd_id && it->cookie == cookie) {
-            it = _mavlink_command_int_handler_table.erase(it);
-        } else {
-            ++it;
-        }
-    }
-
-    // COMMAND_LONG
-    for (auto it = _mavlink_command_long_handler_table.begin();
-         it != _mavlink_command_long_handler_table.end();
-         /* no ++it */) {
-        if (it->cmd_id == cmd_id && it->cookie == cookie) {
-            it = _mavlink_command_long_handler_table.erase(it);
-        } else {
-            ++it;
-        }
-    }
-}
-
-void MavlinkCommandReceiver::unregister_all_mavlink_command_handlers(const void* cookie)
-{
-    std::lock_guard<std::mutex> lock(_mavlink_command_handler_table_mutex);
-
-    // COMMAND_INT
-    for (auto it = _mavlink_command_int_handler_table.begin();
-         it != _mavlink_command_int_handler_table.end();
-         /* no ++it */) {
-        if (it->cookie == cookie) {
-            it = _mavlink_command_int_handler_table.erase(it);
-        } else {
-            ++it;
-        }
-    }
-
-    // COMMAND_LONG
-    for (auto it = _mavlink_command_long_handler_table.begin();
-         it != _mavlink_command_long_handler_table.end();
-         /* no ++it */) {
-        if (it->cookie == cookie) {
-            it = _mavlink_command_long_handler_table.erase(it);
-        } else {
-            ++it;
         }
     }
 }
