+++ src/mavsdk/core/system_impl.h
@@ -3,7 +3,8 @@
 #include "mavlink_address.h"
 #include "mavlink_include.h"
 #include "mavlink_parameters.h"
-#include "mavlink_commands.h"
+#include "mavlink_command_receiver.h"
+#include "mavlink_command_sender.h"
 #include "mavlink_message_handler.h"
 #include "mavlink_mission_transfer.h"
 #include "mavlink_statustext_handler.h"
